small adjustments from my new project
I am currently working on a new project that basically works off a code like this one. I just pasted a certain copy of it here. It is almost exactly the same code as what I put here long ago. It just now has a win function, will restart automatically, and is much better at not spawning the food where the snakes tail is.